@@ -1,24 +1,19 @@
 {
-  "author": "Thierry Wong and Burnsred",
+  "author": "BurnsRED",
   "license": "MIT",
+  "type": "module",
   "name": "burnsred-entity-library",
   "private": true,
   "repository": "https://github.com/burnsred/burnsred-entity-library.git",
-  "version": "v0.6.2",
+  "version": "0.6.2",
   "workspaces": [
     "packages/*"
   ],
   "engines": {
-    "node": "10.15.3",
-    "npm": "6.4.1"
-  },
-  "jestWebpackResolver": {
-    "silent": true,
-    "webpackConfig": "./configs/webpack.config.js"
+    "node": "16.16.0",
+    "npm": "8.11.0"
   },
   "scripts": {
-    "check": "npm run check:types",
-    "check:types": "tsc",
     "clean": "lerna clean",
     "coveralls": "jest --config configs/jest.config.js --coverage --collectCoverageFrom=packages/**/*.{js,jsx} --collectCoverageFrom=!packages/**/*stories.{js,jsx} --maxWorkers=2 && cat ./coverage/lcov.info | coveralls",
     "deploy": "./scripts/deploy-ghpages.sh",
@@ -28,85 +23,53 @@
     "lerna": "lerna",
     "lint": "npm run lint:js && npm run lint:css",
     "lint:css": "stylelint '**/*.{js,jsx,ts,tsx}'",
-    "lint:js": "eslint --ext .js,.jsx,.ts,.tsx .",
-    "packages-build": "lerna run --scope '@burnsred/*' build",
-    "packages-destroy": "lerna run --scope '@burnsred/*' destroy",
-    "packages-link-main:dist": "lerna run --scope '@burnsred/*' link-main:dist",
-    "packages-link-main:src": "lerna run --scope '@burnsred/*' link-main:src",
-    "packages-link-module:dist": "lerna run --scope '@burnsred/*' link-module:dist",
-    "packages-link-module:src": "lerna run --scope '@burnsred/*' link-module:src",
+    "lint:js": "eslint .",
     "packages-publish": "./scripts/publish.sh",
     "test": "jest --config configs/jest.config.js",
     "yarn": "yarn"
   },
   "dependencies": {
-    "@burnsred/default": "0.6.1",
-    "lodash": "4.17.11",
+    "@burnsred/default": "0.6.2",
+    "core-js": "^3.25.5",
     "prop-types": "15.7.2",
     "react": "16.8.4",
+    "react-dom": "16.8.4",
     "react-intl": "2.8.0"
   },
   "devDependencies": {
-    "@babel/core": "7.3.4",
-    "@babel/plugin-proposal-class-properties": "7.3.4",
-    "@babel/plugin-proposal-export-default-from": "7.2.0",
-    "@babel/plugin-proposal-export-namespace-from": "7.2.0",
-    "@babel/plugin-proposal-optional-chaining": "7.2.0",
-    "@babel/plugin-syntax-dynamic-import": "7.2.0",
-    "@babel/plugin-transform-react-jsx": "7.3.0",
-    "@babel/polyfill": "7.2.5",
-    "@babel/preset-env": "7.3.4",
-    "@babel/preset-react": "7.0.0",
-    "@babel/preset-typescript": "7.3.3",
-    "@types/jest": "24.0.11",
-    "@typescript-eslint/eslint-plugin": "1.4.2",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "10.0.1",
-    "babel-jest": "24.5.0",
-    "babel-loader": "8.0.5",
-    "babel-plugin-idx": "2.4.0",
+    "@babel/core": "7.19.3",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-proposal-export-default-from": "7.18.10",
+    "@babel/plugin-transform-react-jsx": "7.19.0",
+    "@babel/preset-env": "7.19.4",
+    "@babel/preset-react": "7.18.6",
+    "@types/jest": "29.1.2",
+    "babel-jest": "29.1.2",
     "babel-plugin-lodash": "3.3.4",
-    "babel-plugin-react-flow-props-to-prop-types": "0.15.0",
-    "babel-plugin-react-intl-auto": "1.7.0",
-    "babel-plugin-styled-components": "1.10.0",
-    "babel-plugin-transform-react-remove-prop-types": "0.4.24",
-    "coveralls": "3.0.3",
-    "directory-named-webpack-plugin": "4.0.1",
-    "eslint": "5.15.3",
-    "eslint-config-airbnb": "17.1.0",
-    "eslint-config-airbnb-typescript": "1.1.0",
-    "eslint-import-resolver-webpack": "0.11.0",
-    "eslint-plugin-eslint-comments": "3.1.1",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-jest": "22.4.1",
-    "eslint-plugin-jsx-a11y": "6.2.1",
-    "eslint-plugin-promise": "4.0.1",
-    "eslint-plugin-react": "7.12.4",
-    "eslint-plugin-react-hooks": "1.5.1",
-    "eslint-plugin-unicorn": "7.1.0",
-    "jest": "24.5.0",
+    "babel-plugin-styled-components": "2.0.7",
+    "coveralls": "^3.1.1",
+    "eslint": "8.25.0",
+    "eslint-config-airbnb": "19.0.4",
+    "eslint-plugin-eslint-comments": "3.2.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jest": "27.1.1",
+    "eslint-plugin-jsdoc": "^39.3.6",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-promise": "6.0.1",
+    "eslint-plugin-react": "7.31.10",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "eslint-plugin-unicorn": "44.0.2",
+    "jest": "29.1.2",
+    "jest-environment-jsdom": "^29.1.2",
     "jest-prop-type-error": "1.1.0",
-    "jest-styled-components": "6.3.1",
-    "jest-webpack-resolver": "0.3.0",
-    "json": "9.0.6",
+    "jest-styled-components": "7.1.1",
     "lerna": "3.13.1",
-    "lodash": "4.17.11",
     "react-test-renderer": "16.8.4",
     "rimraf": "2.6.3",
-    "rollup": "1.6.0",
-    "rollup-plugin-babel": "4.3.2",
-    "rollup-plugin-copy": "0.2.3",
-    "rollup-plugin-json": "3.1.0",
-    "rollup-plugin-named-directory": "1.0.0",
-    "rollup-plugin-node-resolve": "4.0.1",
-    "rollup-plugin-replace": "2.1.1",
-    "rollup-plugin-terser": "4.0.4",
-    "stylelint": "9.10.1",
-    "stylelint-config-recommended": "2.1.0",
+    "stylelint": "14.13.0",
+    "stylelint-config-recommended": "9.0.0",
     "stylelint-config-styled-components": "0.1.1",
-    "stylelint-processor-styled-components": "1.5.2",
-    "typescript": "3.3.3333",
-    "webpack": "4.29.6",
+    "stylelint-processor-styled-components": "1.10.0",
     "yarn": "1.13.0"
   }
 }
