@@ -1,1 +1,1 @@
-module.exports = 'test-file-stub';
+export default 'test-file-stub';
