@@ -3,9 +3,24 @@ import React from 'react';
 
 import { DefaultContext } from './context';
 
-export default function (mapDefault, defaults = {}) {
+/**
+ * Allow resolving defaults from a shared or supplied context.
+ *
+ * If the name exists in the context, use it.
+ * Else, try to find one of the supplied fallback keys.
+ * If none of them exist, return `undefined`.
+ *
+ * If no map is supplied, it will resolve directly to the `context`.
+ *
+ * @param {object} mapDefault - Map of key => [fallback, ...]
+ * @param {object} [defaults] - override context
+ *
+ * @returns {object}
+ */
+export default function useDefault(mapDefault, defaults = {}) {
   const context = React.useContext(DefaultContext) || defaults;
 
+  // If there are no mappings we can bail out immediately.
   if (!mapDefault) return context;
 
   if (process.env.NODE_ENV === 'production') {
@@ -13,6 +28,8 @@ export default function (mapDefault, defaults = {}) {
     if (!_.isObjectLike(mapDefault)) throw new Error('default: mapDefault must be an object');
   }
 
+  // TODO: Conditional rendering of hooks is not allowed -- this should somehow
+  // be moved above the shortcut exist above.
   return React.useMemo(
     () => _.mapValues(
       mapDefault,
@@ -30,6 +47,8 @@ export default function (mapDefault, defaults = {}) {
         return computedValue && context[computedValue];
       },
     ),
+    // XXX This appears to be a hack so the Memo is re-run when `props` change,
+    // assuming we're passed a components `props`.
     [context, mapDefault, ..._.map(mapDefault, (val, key) => defaults[key])],
   );
 }
