@@ -3,10 +3,28 @@ import React from 'react';
 
 import useDefault from './use-default';
 
-const getDisplayName = ComposedComponent => ComposedComponent.displayName
-  || ComposedComponent.name
-  || 'Component';
+/**
+ * Helper function to resolve the name of a React component.
+ *
+ * @param {React.Component} ComposedComponent
+ *
+ * @returns {string} - Display name for the component.
+ */
+function getDisplayName(ComposedComponent) {
+  return ComposedComponent.displayName
+    || ComposedComponent.name
+    || 'Component';
+}
 
+/**
+ * HOC for applying `useDefault` to a components props
+ *
+ * @param {Function| object} configs
+ * As a Function, it is applied to the `props` to provide the `defaultMap` for `useDefault`
+ * Otherwise, passed directyle as `defaultMap`.
+ *
+ * @returns {React.Component}
+ */
 export default configs => (ComposedComponent) => {
   function withDefault(props) {
     const defaults = useDefault(_.isFunction(configs) ? configs(props) : configs, props); // eslint-disable-line react-hooks/rules-of-hooks
