@@ -1,11 +1,21 @@
-import _ from 'lodash';
-
-export default (value, { field }) => {
+/**
+ * Cleaner to trim a value to the maxLength specified on the field.
+ *
+ * If maxLength is not specified, this is a no-op.
+ *
+ * @param {string} value
+ * @param {object} options
+ * @param {module:field~Field} options.field
+ *
+ * @returns {string}
+ */
+export default function sliceToMaxLength(value, { field }) {
   if (process.env.NODE_ENV !== 'production') {
-    if (!_.isString(value)) throw new Error('cleaners.removeMultiSpace: value must be of type string');
+    // eslint-disable-next-line unicorn/no-lonely-if
+    if (typeof value !== 'string') throw new Error('cleaners.removeMultiSpace: value must be of type string');
   }
 
   return field.maxLength
     ? value.slice(0, field.maxLength)
     : value;
-};
+}
