@@ -1,4 +1,3 @@
-import _ from 'lodash';
 import idx from 'idx';
 import AnyField from './field-any';
 
@@ -7,12 +6,13 @@ export default class IdField extends AnyField {
     super(configs);
 
     if (process.env.NODE_ENV !== 'production') {
+      // eslint-disable-next-line unicorn/no-lonely-if
       if (configs.many) throw new Error(`${this.constructor.name}.constructor: "many" option is not supported.`);
     }
   }
 
   dataToValue(data) {
-    return _.isFunction(idx(data, x => x.toString))
+    return (idx(data, x => x.toString) instanceof Function)
       ? data.toString()
       : data;
   }
