@@ -1,1 +1,6 @@
-export default () => false;
+/**
+ * Dummy validator
+ *
+ * @type {module:validator~fieldValidatorFunc}
+ */
+export default function allowBlank() { return false; }
