@@ -1,7 +1,18 @@
 import isRequired from './is-required';
 
-export default (value, configs = {}) => {
+/**
+ * Validator requiring the value not be what the Field considers blank.
+ *
+ * @type {module:validator~fieldValidatorFunc}
+ *
+ * @param {any} value
+ * @param {object} configs
+ * @param {module:field~Field} configs.field
+ * @param {string} [configs.flag]
+ */
+export default function mayNotBeBlank(value, configs = {}) {
   if (process.env.NODE_ENV !== 'production') {
+    // eslint-disable-next-line unicorn/no-lonely-if
     if (!configs.field) throw new Error('validator.mayNotBeBlank: "field" option is required');
   }
 
@@ -17,4 +28,4 @@ export default (value, configs = {}) => {
   return defaultError
     ? flagError && defaultError
     : flagError && 'May not be blank';
-};
+}
